Add ability to do follow-up queries of revisions.
This is added through the exported function GET-REVISIONS-RESULT,
which returns a QUERY-RESULT object.

A QUERY-RESULT object accumulates results from a query,
such as GET-REVISIONS, that may require follow up queries to
retrieve more results. The methods HAS-MORE-RESULTS-P and
GET-MORE-RESULTS-P can be used to continue the query, and
RESULTS returns the results.

Results are now stored as a list of alists. I 
probably will update this to store results as vectors.

QUERY-RESULT and friends have been written generically so they
can be used to add similar behaviour to other queries besides
GET-REVISIONS. The non-exported functions
GET-REVISIONS-AND-CLOSURE does most of the
magic for follow-up queries.