added set-section-content, which accepts a page, rvsection, and text
Uses rvsection as returned by list-page-sections. One funky bit I'm not
fond of: the new text must contain the markup for the section header. So
if you use list-page-sections to figure out that "== foo ==" is rvsection 4,
then when using set-section-content your text has to start with "== foo =="
or else the section will be lost. If the new text does not start with
"== foo ==", the content will be saved as the end of rvsection 3, and
rvsection 4 will now refer to a different section.

Right now we throw an error if the text doesn't start with "==", which I
think is the least surprising way to handle it.