Have WITH-MEDIAWIKI properly evaluate its argument
Adds a function that checks the type of the passed object at runtime
rather than at macroexpansion time and then modifies the macro to
generate a call to this function.  This way, if the url of the mediawiki
site is being computed, it won't misinterpret the form that generates
the url as the instance of mediawiki to be worked with.